pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-git-versioning"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "tomato"
authors = [
    {name = "Peter Kraus"},
    {name = "Graham Kimbell"},
]
maintainers = [
    {name = "Peter Kraus", email = "peter.kraus@tu-berlin.de"},
]
description = "au-tomation without pain"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
requires-python = ">= 3.10"
dependencies = [
    "appdirs >= 1.4.0",
    "toml >= 0.10",
    "pyyaml >= 6.0",
    "psutil >= 5.9",
    "dgbowl_schemas >= 119",
    "pyzmq >= 25.1",
    "h5netcdf >= 1.3",
    "xarray >= 2024.10.0",
    "pydantic >= 2.0",
]

[project.optional-dependencies]
testing = [
    "pytest",
    # "tomato-example-counter >= 1.0.0",
    "tomato-example-counter @ git+https://github.com/dgbowl/tomato-example-counter.git@driver_2.0",
    "tomato-psutil >= 1.0.0",
]
docs = [
    "sphinx ~= 7.2",
    "sphinx-rtd-theme ~= 1.3.0",
    "autodoc-pydantic ~= 2.1",
    "sphinxcontrib-mermaid ~= 0.9.2",
]

[project.urls]
Documentation = "https://dgbowl.github.io/tomato/main"
Issues = "https://github.com/dgbowl/tomato/issues"
Repository = "https://github.com/dgbowl/tomato/"

[project.scripts]
tomato = "tomato:run_tomato"
ketchup = "tomato:run_ketchup"
passata = "tomato:run_passata"
tomato-job = "tomato.daemon.job:tomato_job"
tomato-driver = "tomato.daemon.driver:tomato_driver"
tomato-daemon = "tomato.daemon:tomato_daemon"

[tool.setuptools-git-versioning]
enabled = true
dev_template = "{tag}.post{ccount}"
dirty_template = "{tag}.post{ccount}+dirty"

[tool.ruff]

[tool.pytest.ini_options]
# log_cli = true
# log_cli_level = "DEBUG"